+++ packages/store/src/decorators/select.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
 
-import { Store } from './store';
-import { fastPropGetter } from './internals';
-import { META_KEY } from './symbols';
+import { Store } from '../store';
+import { fastPropGetter } from '../internal/internals';
+import { META_KEY } from '../symbols';
 
 /**
  * Allows the select decorator to get access to the DI store.
