+++ packages/store/src/internal/state-stream.ts
