+++ packages/store/src/plugin_api.ts
@@ -1,5 +1,5 @@
 export { NgxsModule } from './module';
 export { NGXS_PLUGINS, NgxsPlugin } from './symbols';
-export { StateStream } from './state-stream';
-export { getActionTypeFromInstance, setValue, getValue } from './utils';
-export { InitState, UpdateState } from './actions';
+export { StateStream } from './internal/state-stream';
+export { getActionTypeFromInstance, setValue, getValue } from './utils/utils';
+export { InitState, UpdateState } from './actions/actions';
