+++ packages/store/src/public_api.ts
@@ -1,10 +1,10 @@
 export { NgxsModule } from './module';
-export { Action } from './action';
+export { Action } from './decorators/action';
 export { Store } from './store';
-export { State } from './state';
-export { Select } from './select';
+export { State } from './decorators/state';
+export { Select } from './decorators/select';
 export { Actions } from './actions-stream';
-export { ofAction, ofActionSuccessful, ofActionDispatched, ofActionErrored } from './of-action';
+export { ofAction, ofActionSuccessful, ofActionDispatched, ofActionErrored } from './operators/of-action';
 export { NgxsPlugin, NgxsPluginFn, StateContext, NgxsOnInit } from './symbols';
-export { Selector } from './selector';
-export { getActionTypeFromInstance, actionMatcher } from './utils';
+export { Selector } from './decorators/selector';
+export { getActionTypeFromInstance, actionMatcher } from './utils/utils';
