+++ packages/store/src/utils/freeze.ts
