+++ packages/store/tests/module.spec.ts
@@ -2,9 +2,9 @@ import { NgModule } from '@angular/core';
 import { async, TestBed } from '@angular/core/testing';
 
 import { NgxsModule } from '../src/module';
-import { State } from '../src/state';
+import { State } from '../src/decorators/state';
 import { Store } from '../src/store';
-import { Selector } from '../src/selector';
+import { Selector } from '../src/decorators/selector';
 
 interface RootStateModel {
   foo: string;
